incorporate PR review suggestions (javadoc cleanup, renames, add missing annotations
