+++ geode-assembly/src/integrationTest/resources/assembly_content.txt
@@ -747,7 +747,7 @@ javadoc/org/apache/geode/management/package-tree.html
 javadoc/org/apache/geode/management/runtime/CacheServerInfo.html
 javadoc/org/apache/geode/management/runtime/GatewayReceiverInfo.html
 javadoc/org/apache/geode/management/runtime/MemberInformation.html
-javadoc/org/apache/geode/management/runtime/RebalanceResult.PerRegionStats.html
+javadoc/org/apache/geode/management/runtime/RebalanceRegionResult.html
 javadoc/org/apache/geode/management/runtime/RebalanceResult.html
 javadoc/org/apache/geode/management/runtime/RuntimeInfo.html
 javadoc/org/apache/geode/management/runtime/RuntimeRegionInfo.html
